fix(source-proofs): satisfy catalog and architecture gates
