+++ sources/langfuse/testdata/discover_metric.json
@@ -1,0 +1,3 @@
+[
+  "urn:cerebro:writer:langfuse_metric:example-ff1256f0"
+]
