+++ sources/langfuse/testdata/discover_observation.json
@@ -1,0 +1,3 @@
+[
+  "urn:cerebro:writer:langfuse_observation:example-4c3ed17a"
+]
