+++ sources/langfuse/testdata/discover_project.json
@@ -1,0 +1,3 @@
+[
+  "urn:cerebro:writer:langfuse_project:example-36bcc519"
+]
