+++ sources/writer/testdata/discover_application.json
@@ -1,0 +1,3 @@
+[
+  "urn:cerebro:writer:writer_application:example-3082e709"
+]
