+++ sources/writer/testdata/discover_application_job.json
@@ -1,0 +1,1 @@
+[]
