+++ sources/writer/testdata/discover_file.json
@@ -1,0 +1,3 @@
+[
+  "urn:cerebro:writer:writer_file:example-8628b526"
+]
