+++ sources/writer/testdata/discover_graph.json
@@ -1,0 +1,3 @@
+[
+  "urn:cerebro:writer:writer_graph:example-66500c83"
+]
