+++ sources/writer/testdata/discover_model.json
@@ -1,0 +1,101 @@
+[
+  "urn:cerebro:writer:writer_model:example-6c78e072",
+  "urn:cerebro:writer:writer_model:example-b016adb9",
+  "urn:cerebro:writer:writer_model:example-2e59e91b",
+  "urn:cerebro:writer:writer_model:example-47463e31",
+  "urn:cerebro:writer:writer_model:example-f1e74daf",
+  "urn:cerebro:writer:writer_model:example-46878cbd",
+  "urn:cerebro:writer:writer_model:example-5d79ab31",
+  "urn:cerebro:writer:writer_model:example-283df740",
+  "urn:cerebro:writer:writer_model:example-18a473eb",
+  "urn:cerebro:writer:writer_model:example-c4c23f56",
+  "urn:cerebro:writer:writer_model:example-c6bafb6f",
+  "urn:cerebro:writer:writer_model:example-78746a56",
+  "urn:cerebro:writer:writer_model:example-2e9167f5",
+  "urn:cerebro:writer:writer_model:example-e76fc72c",
+  "urn:cerebro:writer:writer_model:example-556d218a",
+  "urn:cerebro:writer:writer_model:example-13ddbd65",
+  "urn:cerebro:writer:writer_model:example-0bb2fb65",
+  "urn:cerebro:writer:writer_model:example-8f9cf6b8",
+  "urn:cerebro:writer:writer_model:example-10cce38b",
+  "urn:cerebro:writer:writer_model:example-125e6f66",
+  "urn:cerebro:writer:writer_model:example-9ac58ddc",
+  "urn:cerebro:writer:writer_model:example-dee49bcd",
+  "urn:cerebro:writer:writer_model:example-56777882",
+  "urn:cerebro:writer:writer_model:example-17c03c7d",
+  "urn:cerebro:writer:writer_model:example-0b759938",
+  "urn:cerebro:writer:writer_model:example-232f8338",
+  "urn:cerebro:writer:writer_model:example-6d6fdcd1",
+  "urn:cerebro:writer:writer_model:example-9bf9e8d6",
+  "urn:cerebro:writer:writer_model:example-5e07e2dd",
+  "urn:cerebro:writer:writer_model:example-84d7b7ff",
+  "urn:cerebro:writer:writer_model:example-543b2f6f",
+  "urn:cerebro:writer:writer_model:example-aca50385",
+  "urn:cerebro:writer:writer_model:example-64c865bd",
+  "urn:cerebro:writer:writer_model:example-9f1bcab0",
+  "urn:cerebro:writer:writer_model:example-d7c4935e",
+  "urn:cerebro:writer:writer_model:example-da9c977d",
+  "urn:cerebro:writer:writer_model:example-9ad7267a",
+  "urn:cerebro:writer:writer_model:example-b5c998c0",
+  "urn:cerebro:writer:writer_model:example-8a000d7e",
+  "urn:cerebro:writer:writer_model:example-079c8b8e",
+  "urn:cerebro:writer:writer_model:example-fca1e380",
+  "urn:cerebro:writer:writer_model:example-07a8d83a",
+  "urn:cerebro:writer:writer_model:example-bc25b6ee",
+  "urn:cerebro:writer:writer_model:example-8a2f3278",
+  "urn:cerebro:writer:writer_model:example-762202e2",
+  "urn:cerebro:writer:writer_model:example-65711717",
+  "urn:cerebro:writer:writer_model:example-db3e9623",
+  "urn:cerebro:writer:writer_model:example-fab29cf0",
+  "urn:cerebro:writer:writer_model:example-232d2425",
+  "urn:cerebro:writer:writer_model:example-73d627db",
+  "urn:cerebro:writer:writer_model:example-96ab5629",
+  "urn:cerebro:writer:writer_model:example-934e6160",
+  "urn:cerebro:writer:writer_model:example-e71ebb16",
+  "urn:cerebro:writer:writer_model:example-19ecf77a",
+  "urn:cerebro:writer:writer_model:example-d60b7267",
+  "urn:cerebro:writer:writer_model:example-df0dfbef",
+  "urn:cerebro:writer:writer_model:example-7176a2eb",
+  "urn:cerebro:writer:writer_model:example-3a292577",
+  "urn:cerebro:writer:writer_model:example-07b87dee",
+  "urn:cerebro:writer:writer_model:example-26c492b5",
+  "urn:cerebro:writer:writer_model:example-d3e9d049",
+  "urn:cerebro:writer:writer_model:example-433424a4",
+  "urn:cerebro:writer:writer_model:example-e086f286",
+  "urn:cerebro:writer:writer_model:example-fa4b56a4",
+  "urn:cerebro:writer:writer_model:example-570fc318",
+  "urn:cerebro:writer:writer_model:example-9712e4c4",
+  "urn:cerebro:writer:writer_model:example-102810f3",
+  "urn:cerebro:writer:writer_model:example-9ffcced3",
+  "urn:cerebro:writer:writer_model:example-b83835f4",
+  "urn:cerebro:writer:writer_model:example-e6e51ea9",
+  "urn:cerebro:writer:writer_model:example-2309deab",
+  "urn:cerebro:writer:writer_model:example-78fd7b89",
+  "urn:cerebro:writer:writer_model:example-c4c47752",
+  "urn:cerebro:writer:writer_model:example-f3d23c50",
+  "urn:cerebro:writer:writer_model:example-2537bc6e",
+  "urn:cerebro:writer:writer_model:example-5cea73e1",
+  "urn:cerebro:writer:writer_model:example-8a90b162",
+  "urn:cerebro:writer:writer_model:example-465c8d49",
+  "urn:cerebro:writer:writer_model:example-738d9c88",
+  "urn:cerebro:writer:writer_model:example-c2368516",
+  "urn:cerebro:writer:writer_model:example-2bbf1110",
+  "urn:cerebro:writer:writer_model:example-f4656300",
+  "urn:cerebro:writer:writer_model:example-21d08292",
+  "urn:cerebro:writer:writer_model:example-05b960c5",
+  "urn:cerebro:writer:writer_model:example-bf44d8e6",
+  "urn:cerebro:writer:writer_model:example-ad5065df",
+  "urn:cerebro:writer:writer_model:example-56e48c6e",
+  "urn:cerebro:writer:writer_model:example-289d6999",
+  "urn:cerebro:writer:writer_model:example-696224b7",
+  "urn:cerebro:writer:writer_model:example-3ea737e9",
+  "urn:cerebro:writer:writer_model:example-791bd19a",
+  "urn:cerebro:writer:writer_model:example-71e288f6",
+  "urn:cerebro:writer:writer_model:example-2ea3c6a6",
+  "urn:cerebro:writer:writer_model:example-04ce3953",
+  "urn:cerebro:writer:writer_model:example-8f7d7a37",
+  "urn:cerebro:writer:writer_model:example-b863fef3",
+  "urn:cerebro:writer:writer_model:example-8d02d67c",
+  "urn:cerebro:writer:writer_model:example-7fa5da73",
+  "urn:cerebro:writer:writer_model:example-04dd7771"
+]
