+++ sources/writer/testdata/read_application_job.json
@@ -1,0 +1,1 @@
+[]
