+++ sources/writer/testdata/read_file.json
@@ -1,0 +1,31 @@
+[
+  {
+    "id": "writer-writer-file-014fc11ad601043c",
+    "tenant_id": "writer",
+    "source_id": "writer",
+    "kind": "writer.file",
+    "occurred_at": "2026-08-20T19:30:44Z",
+    "schema_ref": "writer/file/v1",
+    "payload": {
+      "created_at": "2026-08-20T00:24:21.862464Z",
+      "graph_ids": [
+        "example-536aba21"
+      ],
+      "id": "example-8628b526",
+      "name": "example-a4b0f2ae",
+      "status": "failed"
+    },
+    "attributes": {
+      "created_at": "2026-08-20T00:24:21.862464Z",
+      "external_id": "example-8628b526",
+      "family": "file",
+      "file_id": "example-8628b526",
+      "graph_ids": "example-536aba21",
+      "name": "example-a4b0f2ae",
+      "provider": "writer",
+      "source_product": "writer",
+      "source_provider": "writer",
+      "status": "failed"
+    }
+  }
+]
