+++ sources/writer/testdata/read_graph.json
@@ -1,0 +1,39 @@
+[
+  {
+    "id": "writer-writer-graph-a8a9f8ae5bdcffde",
+    "tenant_id": "writer",
+    "source_id": "writer",
+    "kind": "writer.graph",
+    "occurred_at": "2026-08-20T19:30:43Z",
+    "schema_ref": "writer/graph/v1",
+    "payload": {
+      "created_at": "2026-07-15T19:13:19.101910Z",
+      "description": "",
+      "file_status": {
+        "completed": 8,
+        "failed": 0,
+        "in_progress": 0,
+        "total": 8
+      },
+      "id": "example-66500c83",
+      "name": "example-020b8e6e",
+      "type": "manual",
+      "urls": null
+    },
+    "attributes": {
+      "created_at": "2026-07-15T19:13:19.101910Z",
+      "external_id": "example-66500c83",
+      "family": "graph",
+      "file_status_completed": "8",
+      "file_status_failed": "0",
+      "file_status_in_progress": "0",
+      "file_status_total": "8",
+      "graph_id": "example-66500c83",
+      "name": "example-020b8e6e",
+      "provider": "writer",
+      "source_product": "writer",
+      "source_provider": "writer",
+      "type": "manual"
+    }
+  }
+]
